working

+++ internal/database/mobile_recharge.go
@@ -776,7 +776,7 @@ func (db *Database) MobileRechargeRefundQuery(
 	query := `
 		UPDATE mobile_recharge
 		SET status = 'REFUND'
-		WHERE mobile_recharge_transaction_id = @transaction_id::BIGINT;
+		WHERE mobile_recharge_transaction_id = @transaction_id;
 	`
 
 	if _, err := tx.Exec(ctx, query, pgx.NamedArgs{
